requirements.txt
fastapi >= latest
python-multipart >= latest
uvicorn >= latest
python-jose[cryptography] >= latest
passlib[bcrypt] >= latest
fastapi-cache2 >= latest
beanie >= latest